refactor(web): W10 slice 2 — govern/governance collision: policy-decision modules to lib/governance (BI-AB17E1A8)
The pass §3.1-b table flags lib/govern (99 modules) vs lib/governance (11)
as "actively confusing". Contents-based resolution per the W10 brief:
auth/credentials/permissions KEEP the lib/govern name; the four
governance-decision modules misfiled among them move to lib/governance,
which is the governance-plane home (finding contract, severity ladder,
compliance evidence, governance hooks):

- governance-data.ts (+test), governance-resolver.ts (+test),
  governance-types.ts, user-governance.ts -> lib/governance/

Move-only; mechanical follow-through:
- 4 legacy root shims (lib/governance-*.ts, lib/user-governance.ts)
  re-point at ./governance/*;
- lib/govern/principal-context.ts relative import follows
  (../governance/governance-types);
- lib/govern barrel drops the moved exports (barrel has zero external
  importers; its self-test untouched — it never covered the moved four);
- lib/optimization/consolidation-bets.ts file-path selector updated.

Characterization: lib/govern + lib/governance vitest 52 files / 517 tests
green; per-file test-count diff vs the pre-move baseline is EMPTY modulo
the two renamed test paths. Application-boundary guard unchanged:
8 contexts, 531 cross-context imports, 63 exceptions.

Docs-Impact-Decision: internal namespace hygiene, no user-facing change.
Design-Grounding-Decision: grounded in
docs/architecture/2026-08-16-simplify-and-strengthen-architecture-pass.md
§3.1-b + §5 W10 (BI-AB17E1A8); substrate reviewed — lib/governance
finding-contract/severity plane (EP-8DC217EB BET-12) as destination,
root shim layer (Phase 13 refactoring) as the compatibility surface,
lib/govern barrel importer sweep (none).
Process-Spine-Decision: executes a documented architecture-pass move
(§3.1-b govern/governance collision); mechanical git mv + import rewrites.
