+++ apps/web/lib/governance-data.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/govern/governance-data.ts (Phase 13 refactoring)
-export * from "./govern/governance-data";
+// Shim — moved to lib/governance/governance-data.ts (Phase 13 refactoring)
+export * from "./governance/governance-data";
