+++ apps/web/lib/governance-resolver.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/govern/governance-resolver.ts (Phase 13 refactoring)
-export * from "./govern/governance-resolver";
+// Shim — moved to lib/governance/governance-resolver.ts (Phase 13 refactoring)
+export * from "./governance/governance-resolver";
