+++ apps/web/lib/governance-types.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/govern/governance-types.ts (Phase 13 refactoring)
-export * from "./govern/governance-types";
+// Shim — moved to lib/governance/governance-types.ts (Phase 13 refactoring)
+export * from "./governance/governance-types";
