+++ apps/web/lib/user-governance.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/govern/user-governance.ts (Phase 13 refactoring)
-export * from "./govern/user-governance";
+// Shim — moved to lib/governance/user-governance.ts (Phase 13 refactoring)
+export * from "./governance/user-governance";
